host instegration tests + chaos
